Fixed device-management-inventory-internal module package names
