+++ service/device/management/inventory/internal/src/main/java/org/eclipse/kapua/service/device/management/inventory/internal/DeviceInventoryManagementFactoryImpl.java
@@ -10,18 +10,14 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.service.device.management.bundle.internal;
+package org.eclipse.kapua.service.device.management.inventory.internal;
 
 import org.eclipse.kapua.locator.KapuaProvider;
-import org.eclipse.kapua.service.device.management.bundle.model.bundle.internal.internal.DeviceInventoryBundleImpl;
-import org.eclipse.kapua.service.device.management.bundle.model.bundle.internal.internal.DeviceInventoryBundlesImpl;
 import org.eclipse.kapua.service.device.management.bundle.model.inventory.internal.DeviceInventoryImpl;
 import org.eclipse.kapua.service.device.management.bundle.model.inventory.internal.DeviceInventoryItemImpl;
-import org.eclipse.kapua.service.device.management.bundle.model.inventory.packages.internal.DeviceInventoryPackageImpl;
-import org.eclipse.kapua.service.device.management.bundle.model.inventory.packages.internal.DeviceInventoryPackagesImpl;
-import org.eclipse.kapua.service.device.management.bundle.model.inventory.system.internal.DeviceInventorySystemPackageImpl;
-import org.eclipse.kapua.service.device.management.bundle.model.inventory.system.internal.DeviceInventorySystemPackagesImpl;
 import org.eclipse.kapua.service.device.management.inventory.DeviceInventoryManagementFactory;
+import org.eclipse.kapua.service.device.management.inventory.model.bundle.internal.DeviceInventoryBundleImpl;
+import org.eclipse.kapua.service.device.management.inventory.model.bundle.internal.DeviceInventoryBundlesImpl;
 import org.eclipse.kapua.service.device.management.inventory.model.bundle.inventory.DeviceInventoryBundle;
 import org.eclipse.kapua.service.device.management.inventory.model.bundle.inventory.DeviceInventoryBundles;
 import org.eclipse.kapua.service.device.management.inventory.model.inventory.DeviceInventory;
@@ -30,6 +26,10 @@ import org.eclipse.kapua.service.device.management.inventory.model.inventory.pac
 import org.eclipse.kapua.service.device.management.inventory.model.inventory.packages.DeviceInventoryPackages;
 import org.eclipse.kapua.service.device.management.inventory.model.inventory.system.DeviceInventorySystemPackage;
 import org.eclipse.kapua.service.device.management.inventory.model.inventory.system.DeviceInventorySystemPackages;
+import org.eclipse.kapua.service.device.management.inventory.model.packages.internal.DeviceInventoryPackageImpl;
+import org.eclipse.kapua.service.device.management.inventory.model.packages.internal.DeviceInventoryPackagesImpl;
+import org.eclipse.kapua.service.device.management.inventory.model.system.internal.DeviceInventorySystemPackageImpl;
+import org.eclipse.kapua.service.device.management.inventory.model.system.internal.DeviceInventorySystemPackagesImpl;
 
 /**
  * {@link DeviceInventoryManagementFactory} implementation.
