+++ service/device/management/inventory/internal/src/main/java/org/eclipse/kapua/service/device/management/inventory/internal/exception/InventoryManagementResponseException.java
@@ -10,30 +10,34 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.service.device.management.bundle.internal.exception;
+package org.eclipse.kapua.service.device.management.inventory.internal.exception;
 
+import org.eclipse.kapua.service.device.management.exception.DeviceManagementResponseException;
+import org.eclipse.kapua.service.device.management.inventory.DeviceInventoryManagementService;
 import org.eclipse.kapua.service.device.management.message.response.KapuaResponseCode;
 import org.eclipse.kapua.service.device.management.message.response.KapuaResponseMessage;
 import org.eclipse.kapua.service.device.management.message.response.KapuaResponsePayload;
 
 /**
- * {@link InventoryManagementResponseException} to {@code throw} when the {@link KapuaResponseMessage#getResponseCode()} is not {@link KapuaResponseCode#ACCEPTED}.
+ * Base {@link DeviceManagementResponseException} for {@link DeviceInventoryManagementService}.
  *
  * @since 1.5.0
  */
-public class DeviceInventoryGetManagementException extends InventoryManagementResponseException {
+public abstract class InventoryManagementResponseException extends DeviceManagementResponseException {
 
-    private static final long serialVersionUID = -2114589401530177749L;
+    private static final long serialVersionUID = -146419005237479192L;
 
     /**
      * Constructor.
      *
+     * @param code             The {@link InventoryManagementResponseErrorCodes}.
      * @param responseCode     The {@link KapuaResponseMessage#getResponseCode()}
      * @param exceptionMessage The {@link KapuaResponsePayload#getExceptionMessage()}
      * @param exceptionStack   The {@link KapuaResponsePayload#getExceptionStack()}
+     * @param arguments        The additional argument associated with the {@link InventoryManagementResponseException}.
      * @since 1.5.0
      */
-    public DeviceInventoryGetManagementException(KapuaResponseCode responseCode, String exceptionMessage, String exceptionStack) {
-        super(InventoryManagementResponseErrorCodes.INVENTORY_GET_ERROR, responseCode, exceptionMessage, exceptionStack);
+    public InventoryManagementResponseException(InventoryManagementResponseErrorCodes code, KapuaResponseCode responseCode, String exceptionMessage, String exceptionStack, Object... arguments) {
+        super(code, responseCode, exceptionMessage, exceptionStack, arguments);
     }
 }
