+++ service/device/management/inventory/internal/src/main/java/org/eclipse/kapua/service/device/management/inventory/message/internal/InventoryRequestMessage.java
