+++ service/device/management/inventory/internal/src/main/java/org/eclipse/kapua/service/device/management/inventory/message/internal/InventoryResponseChannel.java
@@ -12,8 +12,8 @@
  *******************************************************************************/
 package org.eclipse.kapua.service.device.management.bundle.message.internal;
 
-import org.eclipse.kapua.service.device.management.bundle.internal.DeviceInventoryAppProperties;
 import org.eclipse.kapua.service.device.management.commons.message.response.KapuaResponseChannelImpl;
+import org.eclipse.kapua.service.device.management.inventory.internal.DeviceInventoryAppProperties;
 import org.eclipse.kapua.service.device.management.inventory.model.inventory.DeviceInventory;
 import org.eclipse.kapua.service.device.management.message.response.KapuaResponseChannel;
